First attempt to seperate more stats and cdr away from the core running code
