+++ daemon/call.h
@@ -19,6 +19,7 @@
 #include "socket.h"
 #include "media_socket.h"
 #include "recording.h"
+#include "statistics.h"
 
 #define UNDEFINED ((unsigned int) -1)
 #define TRUNCATED " ... Output truncated. Increase Output Buffer ...                                    \n"
@@ -109,6 +110,7 @@ enum call_type {
 #endif
 
 #define IS_FOREIGN_CALL(c) (c->foreign_call)
+#define IS_OWN_CALL(c) !IS_FOREIGN_CALL(c)
 
 /* flags shared by several of the structs below */
 #define SHARED_FLAG_IMPLICIT_RTCP		0x00000001
@@ -232,54 +234,6 @@ struct transport_protocol {
 };
 extern const struct transport_protocol transport_protocols[];
 
-struct stats {
-	atomic64			packets;
-	atomic64			bytes;
-	atomic64			errors;
-	u_int64_t			delay_min;
-	u_int64_t			delay_avg;
-	u_int64_t			delay_max;
-	u_int8_t			in_tos_tclass; /* XXX shouldn't be here - not stats */
-	atomic64			foreign_sessions; // unresponsible via redis notification
-};
-
-struct request_time {
-	mutex_t lock;
-	u_int64_t count;
-	struct timeval time_min, time_max, time_avg;
-};
-
-struct totalstats {
-	time_t 			started;
-	atomic64		total_timeout_sess;
-	atomic64		total_foreign_sessions;
-	atomic64		total_rejected_sess;
-	atomic64		total_silent_timeout_sess;
-	atomic64		total_final_timeout_sess;
-	atomic64		total_regular_term_sess;
-	atomic64		total_forced_term_sess;
-	atomic64		total_relayed_packets;
-	atomic64		total_relayed_errors;
-	atomic64		total_nopacket_relayed_sess;
-	atomic64		total_oneway_stream_sess;
-
-	u_int64_t               foreign_sessions;
-	u_int64_t               own_sessions;
-	u_int64_t               total_sessions;
-
-	mutex_t			total_average_lock; /* for these two below */
-	u_int64_t		total_managed_sess;
-	struct timeval		total_average_call_dur;
-
-	mutex_t			managed_sess_lock; /* for these below */
-	u_int64_t		managed_sess_max; /* per graphite interval statistic */
-	u_int64_t		managed_sess_min; /* per graphite interval statistic */
-
-	mutex_t			total_calls_duration_lock; /* for these two below */
-	struct timeval		total_calls_duration_interval;
-
-	struct request_time	offer, answer, delete;
-};
 
 struct stream_params {
 	unsigned int		index; /* starting with 1 */
@@ -314,14 +268,7 @@ struct loop_protector {
 	unsigned char		buf[RTP_LOOP_PROTECT];
 };
 
-struct rtp_stats {
-	unsigned int		payload_type;
-	atomic64		packets;
-	atomic64		bytes;
-	atomic64		kernel_packets;
-	atomic64		kernel_bytes;
-	atomic64		in_tos_tclass;
-};
+
 
 struct packet_stream {
 	mutex_t			in_lock,
@@ -502,13 +449,6 @@ struct callmaster {
 	struct homer_sender	*homer;
 };
 
-struct call_stats {
-	time_t		last_packet;
-	struct stats	totals[4]; /* rtp in, rtcp in, rtp out, rtcp out */
-};
-
-
-
 struct callmaster *callmaster_new(struct poller *);
 void callmaster_get_all_calls(struct callmaster *m, GQueue *q);
 struct timeval add_ongoing_calls_dur_in_interval(struct callmaster *m,
@@ -548,12 +488,10 @@ void __rtp_stats_update(GHashTable *dst, GHashTable *src);
 const char *get_tag_type_text(enum tag_type t);
 const char *get_opmode_text(enum call_opmode);
 
-
+const struct rtp_payload_type *__rtp_stats_codec(struct call_media *m);
 
 #include "str.h"
 #include "rtp.h"
-
-
 
 INLINE void *call_malloc(struct call *c, size_t l) {
 	void *ret;
