+++ daemon/cdr.h
@@ -1,0 +1,15 @@
+/*
+ * cdr.h
+ *
+ *  Created on: Mar 14, 2017
+ *      Author: fmetz
+ */
+
+#ifndef CDR_H_
+#define CDR_H_
+
+struct call;
+
+void cdr_update_entry(struct call* c);
+
+#endif /* CDR_H_ */
